@@ -7,18 +7,22 @@
     "type": "git",
     "url": "https://github.com/tylermcginnis/react-router-firebase-auth"
   },
+  "scripts": {
+    "start": "webpack-dev-server --progress --colors"
+  },
   "author": "Tyler McGinnis",
   "dependencies": {
-    "firebase": "^2.3.2",
-    "react": "^0.14.6",
-    "react-dom": "^0.14.6",
-    "react-router": "^2.0.0-rc5"
+    "firebase": "^3.3.0",
+    "react": "^15.3.0",
+    "react-dom": "^15.3.0",
+    "react-router": "^2.6.1"
   },
   "devDependencies": {
-    "babel-core": "^6.4.0",
+    "babel-core": "^6.13.2",
     "babel-loader": "^6.2.4",
-    "babel-preset-es2015": "^6.3.13",
-    "babel-preset-react": "^6.3.13",
-    "webpack": "^1.12.14"
+    "babel-preset-es2015": "^6.13.2",
+    "babel-preset-react": "^6.11.1",
+    "webpack": "^1.13.1",
+    "webpack-dev-server": "^1.14.1"
   }
 }
