@@ -1,11 +1,12 @@
 module.exports = {
-  entry: "./app/App.js",
-  output: {
-    filename: "public/bundle.js"
-  },
-  module: {
-    loaders: [
-      {test: /\.js$/, exclude: /node_modules/, loader: "babel"}
-    ]
-  }
+    entry: "./app/App.js",
+    output: {
+        filename: "public/bundle.js"
+    },
+    devtool: 'source-map',
+    module: {
+        loaders: [
+            {test: /\.js$/, exclude: /node_modules/, loader: "babel"}
+        ]
+    }
 };
